+++ rust/operator-binary/src/main.rs
@@ -1,20 +1,26 @@
 mod restart_controller;
 
-use built_info::PKG_VERSION;
-use clap::{Parser, crate_description, crate_version};
+use std::ops::Deref as _;
+
+use clap::Parser;
 use futures::pin_mut;
 use stackable_operator::{
     CustomResourceExt,
-    cli::{Command, ProductOperatorRun},
+    cli::{Command, ProductOperatorRun, RollingPeriod},
     commons::{
         authentication::AuthenticationClass,
         s3::{S3Bucket, S3Connection},
     },
 };
+use stackable_telemetry::{Tracing, tracing::settings::Settings};
+use tracing::level_filters::LevelFilter;
 
 mod built_info {
     include!(concat!(env!("OUT_DIR"), "/built.rs"));
 }
+
+// TODO (@NickLarsenNZ): Change the variable to `CONSOLE_LOG`
+pub const ENV_VAR_CONSOLE_LOG: &str = "COMMONS_OPERATOR_LOG";
 
 #[derive(Parser)]
 #[clap(about, author)]
@@ -28,28 +34,62 @@ async fn main() -> anyhow::Result<()> {
     let opts = Opts::parse();
     match opts.cmd {
         Command::Crd => {
-            AuthenticationClass::print_yaml_schema(PKG_VERSION)?;
-            S3Connection::print_yaml_schema(PKG_VERSION)?;
-            S3Bucket::print_yaml_schema(PKG_VERSION)?;
+            AuthenticationClass::print_yaml_schema(built_info::PKG_VERSION)?;
+            S3Connection::print_yaml_schema(built_info::PKG_VERSION)?;
+            S3Bucket::print_yaml_schema(built_info::PKG_VERSION)?;
         }
         Command::Run(ProductOperatorRun {
             product_config: _,
             watch_namespace,
-            tracing_target,
+            telemetry_arguments,
             cluster_info_opts,
         }) => {
-            stackable_operator::logging::initialize_logging(
-                "COMMONS_OPERATOR_LOG",
-                "commons",
-                tracing_target,
-            );
-            stackable_operator::utils::print_startup_string(
-                crate_description!(),
-                crate_version!(),
-                built_info::GIT_VERSION,
-                built_info::TARGET,
-                built_info::BUILT_TIME_UTC,
-                built_info::RUSTC_VERSION,
+            let _tracing_guard = Tracing::builder()
+                .service_name("commons-operator")
+                .with_console_output((
+                    ENV_VAR_CONSOLE_LOG,
+                    LevelFilter::INFO,
+                    !telemetry_arguments.no_console_output,
+                ))
+                // NOTE (@NickLarsenNZ): Before stackable-telemetry was used, the log directory was
+                // set via an env: `COMMONS_OPERATOR_LOG_DIRECTORY`.
+                // See: https://github.com/stackabletech/operator-rs/blob/f035997fca85a54238c8de895389cc50b4d421e2/crates/stackable-operator/src/logging/mod.rs#L40
+                // Now it will be `ROLLING_LOGS` (or via `--rolling-logs <DIRECTORY>`).
+                .with_file_output(telemetry_arguments.rolling_logs.map(|log_directory| {
+                    let rotation_period = telemetry_arguments
+                        .rolling_logs_period
+                        .unwrap_or(RollingPeriod::Never)
+                        .deref()
+                        .clone();
+
+                    Settings::builder()
+                        .with_environment_variable(ENV_VAR_CONSOLE_LOG)
+                        .with_default_level(LevelFilter::INFO)
+                        .file_log_settings_builder(log_directory, "tracing-rs.log")
+                        .with_rotation_period(rotation_period)
+                        .build()
+                }))
+                .with_otlp_log_exporter((
+                    "OTLP_LOG",
+                    LevelFilter::DEBUG,
+                    telemetry_arguments.otlp_logs,
+                ))
+                .with_otlp_trace_exporter((
+                    "OTLP_TRACE",
+                    LevelFilter::DEBUG,
+                    telemetry_arguments.otlp_traces,
+                ))
+                .build()
+                .init()?;
+
+            tracing::info!(
+                built_info.pkg_version = built_info::PKG_VERSION,
+                built_info.git_version = built_info::GIT_VERSION,
+                built_info.target = built_info::TARGET,
+                built_info.built_time_utc = built_info::BUILT_TIME_UTC,
+                built_info.rustc_version = built_info::RUSTC_VERSION,
+                "Starting {description}",
+                description = built_info::PKG_DESCRIPTION
             );
 
             let client = stackable_operator::client::initialize_operator(